requirements.txt
aiohappyeyeballs>=2.4.4
aiohttp>=3.11.10
aiosignal>=1.3.2
async-timeout>=5.0.1
attrs>=24.3.0
certifi>=2024.12.14
charset-normalizer>=3.4.0
dataclasses-json>=0.6.7
dnspython>=2.7.0
frozenlist>=1.5.0
idna>=3.10
marshmallow>=3.23.1
meraki>=1.27.0
multidict>=6.1.0
mypy-extensions>=1.0.0
numpy>=2.0.2
packaging>=24.2
pandas>=2.2.3
propcache>=0.2.1
pyhumps>=3.8.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2024.2
requests>=2.32.3
requests-toolbelt>=1.0.0
six>=1.17.0
srvlookup>=3.0.0
typing-inspect>=0.9.0
typing_extensions>=4.12.2
tzdata>=2024.2
urllib3>=2.2.3
wxcadm>=4.3.7
xmltodict>=0.14.2
yarl>=1.18.3
